changes to work buffer cache, fd, wd with modern devices

+++ usr/src/kernel/include/sys/socket.h
@@ -246,11 +246,11 @@ int	getpeername __P((int, struct sockaddr *, int *));
 int	getsockname __P((int, struct sockaddr *, int *));
 int	getsockopt __P((int, int, int, void *, int *));
 int	listen __P((int, int));
-ssize_t	recv __P((int, void *, size_t, int));
+int	recv __P((int, void *, int, int));
 int	recvfrom __P((int, void *, int, int,
 		struct sockaddr *, int *));
 int	recvmsg __P((int, struct msghdr *, int));
-ssize_t	send __P((int, const void *, size_t, int));
+int	send __P((int, const void *, int, int));
 int	sendto __P((int, const void *, int, int, const struct sockaddr *, int));
 int	sendmsg __P((int, const struct msghdr *, int));
 int	setsockopt __P((int, int, int, const void *, int));
